requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
email-validator==2.1.1
pytest==8.2.0
pytest-flask==1.3.0
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
Click==8.1.7
WTForms==3.1.2
blinker==1.8.2
itsdangerous==2.2.0
SQLAlchemy==2.0.30
greenlet==3.0.3
